requirements.txt
django
djangorestframework
markdown
django-filterpip install django-filter